update energy [fix Ae=0.1]

+++ src/main.cpp
@@ -163,9 +163,10 @@ int main(int argc, char* argv[])
             std::cout << count << " particle " << std::endl;
             //output_vd(vd,place);    //output particle properties
             
-            //updateEqtnState(vd);    //calc pressure based on local density <-- UNCOMMENT THIS
-
             updateParticleProperties(vd, place, dt, H, turb, simulation);
+
+            //smooth P
+
             double cfl_temp = (dt/H) * (vd.template getProp<i_velocity>(p)[0]+vd.template getProp<i_velocity>(p)[1]+vd.template getProp<i_velocity>(p)[2]);
             cfl = std::max(abs(cfl_temp),cfl);       //look for max not average. needs to be absolute.  should hold
        
@@ -189,7 +190,7 @@ int main(int argc, char* argv[])
         NN = vd.getCellList(r_cut);
 
         // collect statistics on the configuration
-        //if ((i+1) % simulation.frame == 0)
+        if ((i+1) % simulation.frame == 0)
         {
             // Write the particle position for visualization (Without ghost)
             vd.deleteGhost();
